requirements.txt
# RestAPI
fastapi
uvicorn
python-multipart
jinja2

# Deep Learning & Transformers
torch

# ONNX e inferência
onnx
onnxruntime
transformers  
numpy

# OCR e manipulação de documentos
pytesseract
pillow
python-docx
pymupdf
openpyxl
pandas

# Utilitários
tqdm

# Others
scipy